Refactor safe_extract to its own file
